Initial commit for ConfigurationWizard feature.
The Configuration Wizard provides a GUI for quick set-up and
configuration of settings in OTRS.

Each configuration step has a check and a frontend.

Each configuration has a sysconfig registration.

In the SysConfig you can define a list of groups for which the
configuration step
applies, for instance, 'admin', or 'service desk' and 'management'.

If the permission check completed, or if there are no groups
linked to the configuration step, the step is 'available'.

A configuration step can also be 'actionable'; then it needs
immediate action. This is determined by executing a check-module.

If you log in, or after you install a package, OTRS checks to see
if there are actionable wizard steps for you. If so, it redirects
you to the wizard.

You can also start the wizard by activating it via a menu option.
In that case, all available actions are displayed.

An example of a wizard step might be 'Configure the FQDN of the system'.
It would only be 'actionable' (i.e. require immediate attention) if
the current FQDN is 'yourhost.example.com' (the system default).

So if the FQDN is not set and you log in the system as an administrator,
this check is performed, marked as actionable and the Wizard is
launched.

If the FQDN *is* set, and you log in as an administrator, it is not
marked as Actionable. However, if you'd start the Wizard from your
menu you can still see the config step and change the FQDN from there
if you want.

Another example of a configuration step could be 'change your password',
this could apply for any user and it can be marked as 'actionable'
for instance for new accounts.

And yet another example could be to check if the system is registered.

Please note that some of these configuration options (all?) are basically
duplications of functions that are now in SysConfig somewhere, or
in other places of the code.

Also, eventually, we can move stuff such as AgentInfo to the Wizard.
Also, we might make it available for the Customer Interface.