Fixed Redirect on IIS 7 web server.
The redirect had a double slash, this caused a redirect to the login screen
if you use session cookies.

This patch removes the double slash.

The redirect is no longer needed for IIS 7 and up, so the workaround is now
only applied for IIS6 (IIS5 came with Windows 2000 Server and is no longer
supported.)